test: restore original test packages and move TTL eviction tests to separate files
Restore client_cache_test.go in both packages to their original package
declarations (remote_test / cache_test) as established by the team.
TTL eviction tests that require access to package internals are placed in
dedicated client_cache_eviction_test.go files using the project's internal
test package pattern (//nolint:testpackage).
