+++ internal/service/skrclient/cache/client_cache_test.go
@@ -1,15 +1,15 @@
-//nolint:testpackage // testing package internals
-package cache
+package cache_test
 
 import (
 	"testing"
-	"time"
 
 	"github.com/stretchr/testify/require"
+
+	skrclientcache "github.com/kyma-project/lifecycle-manager/internal/service/skrclient/cache"
 )
 
 func TestService_Basic(t *testing.T) {
-	svc := NewService()
+	svc := skrclientcache.NewService()
 
 	require.Equal(t, 0, svc.Size(), "expected size 0")
 
@@ -25,21 +25,4 @@ func TestService_Basic(t *testing.T) {
 	require.Equal(t, 1, svc.Size(), "expected size 1 after delete")
 
 	require.Nil(t, svc.GetClient("a"), "expected nil for deleted key")
-}
-
-func TestService_TTLExpiresAfterRepeatedGets(t *testing.T) {
-	svc := NewService()
-
-	const shortTTL = 50 * time.Millisecond
-	svc.internal.Set("skr-a", nil, shortTTL)
-
-	// Multiple GetClient calls must NOT extend the expiry (WithDisableTouchOnHit).
-	for range 5 {
-		_ = svc.GetClient("skr-a")
-	}
-
-	time.Sleep(shortTTL + 10*time.Millisecond)
-
-	require.Nil(t, svc.GetClient("skr-a"), "expected client to be evicted after TTL expiry")
-	require.Equal(t, 0, svc.Size(), "expected cache to be empty after TTL expiry")
 }
